+++ draft-packages/form/KaizenDraft/Form/TextAreaField/TextAreaField.tsx
@@ -19,8 +19,11 @@ type Props = {
   defaultValue?: string
   validationMessage?: string
   status?: InputStatus
+  autogrow?: boolean
   description?: React.ReactNode
   textAreaRef?: React.RefObject<HTMLTextAreaElement>
+  variant?: "default" | "prominent"
+  maxLength?: number
   onChange?: (event: React.ChangeEvent<HTMLTextAreaElement>) => any
   onBlur?: (event: React.FocusEvent<HTMLTextAreaElement>) => any
   onFocus?: (event: React.FocusEvent<HTMLTextAreaElement>) => any
@@ -34,9 +37,12 @@ const TextAreaField: React.FunctionComponent<Props> = props => {
     defaultValue,
     validationMessage,
     status,
+    autogrow,
     description,
     inline,
     reversed,
+    variant,
+    maxLength,
     placeholder,
     textAreaRef,
     onChange,
@@ -44,6 +50,19 @@ const TextAreaField: React.FunctionComponent<Props> = props => {
     onFocus,
     rows,
   } = props
+
+  const renderDescription = (position: "top" | "bottom") => {
+    if (!description) return null
+    return (
+      <FieldMessage
+        id={`${id}-field-message`}
+        automationId={`${id}-field-description`}
+        message={description}
+        reversed={reversed}
+        position={position}
+      />
+    )
+  }
 
   return (
     <FieldGroup
@@ -57,7 +76,9 @@ const TextAreaField: React.FunctionComponent<Props> = props => {
         htmlFor={`${id}-field-textarea`}
         labelText={labelText}
         reversed={reversed}
+        variant={variant}
       />
+      {variant === "prominent" && renderDescription("top")}
       <TextArea
         id={`${id}-field-textarea`}
         automationId={`${id}-field-textarea`}
@@ -71,7 +92,9 @@ const TextAreaField: React.FunctionComponent<Props> = props => {
         rows={rows}
         reversed={reversed}
         status={status}
+        autogrow={autogrow}
         textAreaRef={textAreaRef}
+        maxLength={maxLength}
       />
       {validationMessage && (
         <FieldMessage
@@ -82,14 +105,7 @@ const TextAreaField: React.FunctionComponent<Props> = props => {
           reversed={reversed}
         />
       )}
-      {description && (
-        <FieldMessage
-          id={`${id}-field-message`}
-          automationId={`${id}-field-description`}
-          message={description}
-          reversed={reversed}
-        />
-      )}
+      {variant === "default" && renderDescription("bottom")}
     </FieldGroup>
   )
 }