+++ draft-packages/events/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.78](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.77...@kaizen/draft-events@1.2.78) (2020-11-25)
+
+**Note:** Version bump only for package @kaizen/draft-events
+
+
+
+
+
 ## [1.2.77](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.76...@kaizen/draft-events@1.2.77) (2020-11-13)
 
 **Note:** Version bump only for package @kaizen/draft-events