+++ draft-packages/select/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.18...@kaizen/draft-select@1.11.19) (2020-11-26)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
+## [1.11.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.17...@kaizen/draft-select@1.11.18) (2020-11-25)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
+## [1.11.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.16...@kaizen/draft-select@1.11.17) (2020-11-19)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
 ## [1.11.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@1.11.15...@kaizen/draft-select@1.11.16) (2020-11-13)
 
 **Note:** Version bump only for package @kaizen/draft-select